sync (#18)
* opt

* gmp loc trace
jni fix

* Compatibility

* fix waring

* jumpnode

* 5.6 fix

* badge fix

* ci: build GMP plugin for Android+Linux via Epic official containers

* ci: chmod Out world-readable so host runner can upload ue4-owned artifacts

* ci: probe BuildPlugin HostProject structure (prep multi-config builds)

* ci: setup Android toolchain, source-only artifact in dedicated job, auto per-version build-UE<ver> tag

* ci: probe for SetupAndroid location / Android support in image (path was wrong)

* ci: Android is best-effort (no fail if toolchain absent); build Linux-only when Android unavailable

* fix(5.8): silence -Werror,-Wunreachable-code (dead branches/breaks/loop); ci: chown .git back to runner before tagging

* ci: chown only Out to ue4 (not whole checkout) -- keeps .git/source owned by runner, removes tag/artifact permission fixups

* fix: ClassDefaultObject deprecated (5.6+) -> GetDefaultObject(false); keeps no-create semantics for compile-time CDO null path

* fix(5.8): dragonbox unreachable dead ternary branch -> take constant-true branch

* refactor(iwyu): drop monolithic CoreUObject.h from GMPSignalsImpl.h -> UObjectBaseUtility.h (verified UE5.8 Win64 BuildPlugin, warning gone)

* ci: install Android toolchain (5-step: JDK17/cmdline-tools/licenses/sdkmanager/env), best-effort -> fall back to Linux-only on any failure, never fail job

* ci: per-UE-version Android SDK/NDK/build-tools (from engine SetupAndroid.bat: 5.4/5.5=NDK25, 5.6-5.8=NDK27); was hardcoded to 5.6 old NDK25

* ci: pass Android env to ue4 via 'env' prefix (not --preserve-environment, which leaked HOME=/root and broke UAT log dir)

* ci: error-aware docker pull retry (backoff on GHCR rate-limit/network, layer-cache resume; abort fast on auth/not-found)

* ci: keep host Android SDK + .NET in free-disk-space (root disk ~145G has ample room; only prune haskell/tool-cache/large-packages)

* ci: drop no-op 'move Docker to /mnt' step -- runner has no separate /mnt partition (df /mnt == /dev/root); root disk 118G free is plenty

* ci: tag push best-effort -- don't fail job when GITHUB_TOKEN rejects workflow-touching ref (auto-recovers once workflow file stabilizes)

* ci: probe Android code-platform support BEFORE installing SDK (BuildPlugin drops Android via InstalledPlatformInfo if not declared; avoid wasting time on toolchain when engine can't target Android)

* opt(JsonDom): arena-based DOM allocation; MessageTag node preview tweaks

* fix(static-store): use Meyers singleton for per-type slot to avoid vague-linkage lazy-init race; assert Store valid in Flex ConnectImpl

* fix: align Puerts script backend and fix UE version compat

* GMP: AngelScript script backend

* GMP: sluaunreal script backend

* GMP: script-backend codegen base + slua/UnLua static-bind codegen + Build.cs switches

- GMPScriptCodeGen: shared base (tag enumeration, GMP-type-name mapper, EmmyLua annotation builder,
  WriteIfChanged, OnMessageTagSignatureChanged subscription) for all script backends.
- GMPAngelScriptCodeGen: refactored onto the base.
- GMPSluaCodeGen / GMPUnLuaCodeGen: SA EmmyLua annotations (shared GMPMessages.lua) + SB per-tag
  strongly-typed lua_CFunctions (checkValue<T> / UnLua::Get<T>), wrapped in GMP_SLUA_STATIC_BIND /
  GMP_UNLUA_STATIC_BIND.
- GMP.Build.cs: bGMPSluaStaticBind / bGMPUnLuaStaticBind switches (PublicDefinitions, default off).
- SluaSupport.h: GMP_RegisterToSlua fallback hook to generated static binds.

* GMP: AngelScript backend three transparent accel passes

* GMP: propagate script-backend accel to Puerts/UnLua/slua + key-baked real-compile verify

* GMP: UnrealCSharp(C#) backend

* GMP: script message-source tracing for all script backends (Editor)

+++ Plugins/GMP/Source/GMP/Private/GMPProtoSerializerBP.cpp
@@ -770,12 +770,12 @@ namespace Proto
 		else if (Prop->IsA<FInt64Property>())
 		{
 			auto Int64Prop = CastFieldChecked<FInt64Property>(Prop);
-			Int64Prop->SetPropertyValue(Addr, FMath::RandRange(INT64_MIN, INT64_MAX));
+			Int64Prop->SetPropertyValue(Addr, FMath::RandRange((int64)INT64_MIN, (int64)INT64_MAX));
 		}
 		else if (Prop->IsA<FUInt64Property>())
 		{
 			auto UInt64Prop = CastFieldChecked<FUInt64Property>(Prop);
-			UInt64Prop->SetPropertyValue(Addr, (uint64)FMath::RandRange(INT64_MIN, INT64_MAX));
+			UInt64Prop->SetPropertyValue(Addr, (uint64)FMath::RandRange((int64)INT64_MIN, (int64)INT64_MAX));
 		}
 		else if (Prop->IsA<FFloatProperty>())
 		{
